+++ services/ui-src/src/views/CombinedRatesPage/CombinedRatesMeasure.tsx
@@ -1,7 +1,7 @@
 import * as QMR from "components";
 import * as CUI from "@chakra-ui/react";
 import { useParams, Link } from "react-router-dom";
-import { CombinedRateDataSource } from "shared/commonQuestions/CombinedRateDataSource/CombinedRateDataSource";
+import { DataSourceInformationBanner } from "shared/commonQuestions/DataSouceInformationBanner/DataSourceInformationBanner";
 
 interface Props {
   year: string;
@@ -42,7 +42,7 @@ export const CombinedRatesMeasure = ({ year, measureName }: Props) => {
           </Link>
         </CUI.ListItem>
       </CUI.UnorderedList>
-      <CombinedRateDataSource />
+      <DataSourceInformationBanner />
     </QMR.StateLayout>
   );
 };
